@@ -1,10 +1,8 @@
-/*
- * Copyright (c) 2020-2021 Digital Bazaar, Inc. All rights reserved.
+/*!
+ * Copyright (c) 2020-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
-const bedrock = require('bedrock');
-const {asyncHandler} = require('bedrock-express');
+import * as bedrock from '@bedrock/core';
+import {asyncHandler} from '@bedrock/express';
 
 bedrock.events.on('bedrock-express.configure.routes', app => {
   app.get(
