@@ -1,14 +1,12 @@
-/*
- * Copyright (c) 2020-2021 Digital Bazaar, Inc. All rights reserved.
+/*!
+ * Copyright (c) 2020-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import * as bedrock from '@bedrock/core';
+import '@bedrock/express';
+import '@bedrock/https-agent';
+import '@bedrock/server-load-balancer-config';
 
-const bedrock = require('bedrock');
-require('bedrock-express');
-require('bedrock-https-agent');
-require('bedrock-server-load-balancer-config');
+import './mock-http.js';
 
-require('./mock-http');
-
-require('bedrock-test');
+import '@bedrock/test';
 bedrock.start();
