@@ -1,8 +1,8 @@
 
 /*!
- * Copyright (c) 2020-2021 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2020-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import bedrock from 'bedrock';
+import {loggers} from '@bedrock/core';
 
-export default bedrock.loggers.get('app')
+export const logger = loggers.get('app')
   .child('bedrock-server-load-balancer-config');
